requirements.txt
crewai>=0.28.8
crewai-tools>=0.1.6
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
linkup-sdk>=1.0.0
